+++ clang/test/Driver/ppc-cpus.c
@@ -39,3 +39,6 @@
 // RUN: %clang -### -c -target powerpc64 %s -mcpu=405     2>&1 | FileCheck %s --check-prefix=GENERIC
 //
 // GENERIC: "-target-cpu" "ppc64"
+
+// RUN: %clang -### -c --target=powerpc64 %s -march=generic 2>&1 | FileCheck --check-prefix=MARCH %s
+// MARCH: error: unsupported option '-march=' for target 'powerpc64'